Add files via upload
You are a manager of a  Wally's Training Gym and you encourage your trainers to enroll new members.  Input is the trainer's last name and the number of new enrollees.  Output is the number of trainers who have enrolled members in each of three categories:  0-5 new members, 6-10 new members, and 11 to 15 new members. Write an application that allows the user to enter 15 Trainer's names in one array and the number of new members they have enrolled into another array.   Output is to display the number of trainers who are in each category.